+++ src/plone/app/fhirfield/helpers.py
@@ -1,26 +1,21 @@
 # _*_ coding: utf-8 _*_
 import os
-import pkgutil
 import re
-import sys
 import time
-from importlib import import_module
+from urllib.parse import unquote_plus
 
-import six
 from fhirpath.enums import FHIR_VERSION
-from fhirpath.utils import lookup_fhir_class_path
+from fhirpath.utils import lookup_fhir_class
+from fhirpath.utils import reraise
 from plone.api.validation import required_parameters
 from plone.app.fhirfield.compat import _
 from plone.app.fhirfield.compat import json
 from plone.memoize import ram
-from six.moves.urllib.parse import unquote_plus
 from zope.interface import Invalid
 
 from .compat import EMPTY_STRING
 from .compat import NO_VALUE
 from .exc import SearchQueryError
-from .exc import SearchQueryValidationError
-from .variables import FHIR_ES_MAPPINGS_CACHE
 from .variables import FHIR_RESOURCE_LIST  # noqa: F401
 from .variables import FHIR_RESOURCE_MODEL_CACHE  # noqa: F401
 from .variables import FHIR_SEARCH_PARAMETER_SEARCHABLE
@@ -33,63 +28,6 @@ PATH_WITH_DOT_IS = re.compile(r"\.is\([a-z]+\)$", re.I)
 PATH_WITH_DOT_WHERE = re.compile(r"\.where\([a-z]+\=\'[a-z]+\'\)$", re.I)
 
 
-@required_parameters("model_name")
-def search_fhir_model(model_name, cache=True):
-    """ """
-    return lookup_fhir_class_path(model_name, fhir_release=FHIR_VERSION.STU3)
-    global FHIR_RESOURCE_MODEL_CACHE
-    if model_name in FHIR_RESOURCE_MODEL_CACHE.keys() and cache:
-        return "{0}.{1}".format(FHIR_RESOURCE_MODEL_CACHE[model_name], model_name)
-
-    # Trying to get from entire modules
-    from fhir import resources
-
-    for importer, modname, ispkg in pkgutil.walk_packages(
-        resources.__path__, resources.__name__ + ".", onerror=lambda x: None
-    ):
-        if ispkg or len(modname.split(".")) > 3:
-            continue
-
-        mod = import_module(modname)
-        if getattr(mod, model_name, None):
-            FHIR_RESOURCE_MODEL_CACHE[model_name] = modname
-            return "{0}.{1}".format(modname, model_name)
-
-    return None
-
-
-@required_parameters("resource_type")
-def resource_type_str_to_fhir_model(resource_type):
-    """ """
-    dotted_path = search_fhir_model(resource_type)
-    if dotted_path is None:
-        raise SearchQueryValidationError(
-            _("Invalid: `{0}` is not valid resource type!".format(resource_type))
-        )
-
-    return import_string(dotted_path)
-
-
-@required_parameters("dotted_path")
-def import_string(dotted_path):
-    """Shameless hack from django utils, please don't mind!"""
-    try:
-        module_path, class_name = dotted_path.rsplit(".", 1)
-    except (ValueError, AttributeError):
-        msg = "{0} doesn't look like a module path".format(dotted_path)
-        six.reraise(ImportError, ImportError(msg), sys.exc_info()[2])
-
-    module = import_module(module_path)
-
-    try:
-        return getattr(module, class_name)
-    except AttributeError:
-        msg = 'Module "{0}" does not define a "{1}" attribute/class'.format(
-            module_path, class_name
-        )
-        six.reraise(ImportError, ImportError(msg), sys.exc_info()[2])
-
-
 @required_parameters("str_val")
 def parse_json_str(str_val, encoding="utf-8"):
     """ """
@@ -100,11 +38,7 @@ def parse_json_str(str_val, encoding="utf-8"):
     try:
         json_dict = json.loads(str_val, encoding=encoding)
     except ValueError as exc:
-        six.reraise(
-            Invalid,
-            Invalid("Invalid JSON String is provided!\n{0!s}".format(exc)),
-            sys.exc_info()[2],
-        )
+        reraise(Invalid, "Invalid JSON String is provided!\n{0!s}".format(exc))
 
     return json_dict
 
@@ -127,7 +61,7 @@ def validate_index_name(name):
             "example: hospital_resource, patient"
         )
 
-        six.reraise(SearchQueryError, SearchQueryError(msg), sys.exc_info()[2])
+        reraise(SearchQueryError, msg)
 
 
 @ram.cache(
@@ -138,7 +72,7 @@ def fhir_search_path_meta_info(path):
     resource_type = path.split(".")[0]
     properties = path.split(".")[1:]
 
-    model_cls = resource_type_str_to_fhir_model(resource_type)
+    model_cls = lookup_fhir_class(resource_type, FHIR_VERSION["STU3"])
     result = None
     for prop in properties:
 
@@ -243,7 +177,7 @@ def parse_query_string(request, allow_none=False):
         >>> REQUEST['QUERY_STRING']
         'patient=P001&lastUpdated=2018-01-01&lastUpdated=lt2018-09-10'
 
-        >>> from six.moves.urllib.parse import urlencode
+        >>> from urllib.parse import urlencode
         >>> params = [('patient', 'P001'), ('lastUpdated', '2018-01-01'),
                       ('lastUpdated', 'lt2018-09-10')]
         >>> urlencode(params)