+++ src/plone/app/fhirfield/interfaces.py
@@ -6,6 +6,7 @@ from zope import schema as zs
 from zope.interface import Attribute
 from zope.interface import Interface
 from zope.schema.interfaces import IObject
+from fhirspec import FHIR_RELEASES
 
 
 class IFhirResourceModel(Interface):
@@ -31,7 +32,11 @@ class IFhirResource(IObject):
         title=_("FHIR Resource Model from fhirclient"), required=False
     )
     model_interface = zs.DottedName(title=_("FHIR Model Interface"), required=False)
-    fhir_version = zs.TextLine(title="FHIR Version(Release)", required=True)
+    fhir_release = zs.Choice(
+        title="FHIR Version(Release)",
+        values=[r.value for r in FHIR_RELEASES],
+        required=True,
+    )
     index_mapping = JSONField(title=_("Index Mapping"), required=False)
 
     def from_dict(dict_value):