+++ src/plone/app/fhirfield/variables.py
@@ -94,5 +94,3 @@ with open(os.path.join(FHIR_RESOURCE_LIST_DIR, FHIR_VERSION + ".json"), "r") as 
     """ """
     FHIR_RESOURCE_LIST = json.load(f)["resources"]
 
-
-FHIR_ES_MAPPINGS_CACHE = dict()